Mudanças de Lógica de funcionamento, adequação de funcionalidades e adicionando frontend do streamlit no projeto.
