+++ backend/app/core/browser.py
@@ -1,8 +1,11 @@
 from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
 from selenium.common.exceptions import WebDriverException
+import sys
+import os
 
 
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 from app.config.settings import BROWSER_USER_AGENT, BROWSER_TIMEOUT
